Rework network rate limit utils
