+++ library/src/main/java/eu/kanade/tachiyomi/network/interceptor/RateLimitInterceptor.kt
@@ -1,10 +1,9 @@
+@file:Suppress("UnusedReceiverParameter")
+
 package eu.kanade.tachiyomi.network.interceptor
 
-import okhttp3.HttpUrl
 import okhttp3.OkHttpClient
 import java.util.concurrent.TimeUnit
-import kotlin.time.Duration
-import kotlin.time.Duration.Companion.seconds
 
 /**
  * An OkHttp interceptor that handles rate limiting.
@@ -23,89 +22,19 @@ import kotlin.time.Duration.Companion.seconds
  * @param period [Long]   The limiting duration. Defaults to 1.
  * @param unit [TimeUnit] The unit of time for the period. Defaults to seconds.
  */
-@Suppress("unused_parameter", "DeprecatedCallableAddReplaceWith")
-@Deprecated("Use the version with kotlin.time APIs instead.")
+@Deprecated(
+    message = "Use the version with kotlin.time APIs instead.",
+    replaceWith = ReplaceWith(
+        expression = "rateLimit(permits, period.toDuration(unit.toDurationUnit()))",
+        imports = [
+            "mihonx.network.rateLimit",
+            "kotlin.time.toDuration",
+            "kotlin.time.toDurationUnit",
+        ]
+    )
+)
 fun OkHttpClient.Builder.rateLimit(
     permits: Int,
     period: Long = 1,
     unit: TimeUnit = TimeUnit.SECONDS,
-): OkHttpClient.Builder = throw RuntimeException("Stub!")
-
-/**
- * An OkHttp interceptor that enforces rate limiting across all requests.
- *
- * Examples:
- *
- * permits = 5,  period = 1.seconds  =>  5 requests per second
- * permits = 10, period = 2.minutes  =>  10 requests per 2 minutes
- *
- * @since extension-lib 14
- *
- * @param permits [Int]     Number of requests allowed within a period of units.
- * @param period [Duration] The limiting duration. Defaults to 1.seconds.
- */
-@Suppress("unused_parameter")
-fun OkHttpClient.Builder.rateLimit(
-    permits: Int,
-    period: Duration = 1.seconds
-): OkHttpClient.Builder = throw RuntimeException("Stub!")
-
-/**
- * An OkHttp interceptor that handles given url host's rate limiting.
- *
- * Examples:
- *
- * url = "https://api.anime.com", permits = 5, period = 1.seconds =>  5 requests per second to api.anime.com
- * url = "https://imagecdn.anime.com", permits = 10, period = 2.minutes  =>  10 requests per 2 minutes to imagecdn.anime.com
- *
- * @since extension-lib 16
- *
- * @param url [String]      The url host that this interceptor should handle. Will get url's host by using HttpUrl.host()
- * @param permits [Int]     Number of requests allowed within a period of units.
- * @param period [Duration] The limiting duration. Defaults to 1.seconds.
- */
-fun OkHttpClient.Builder.rateLimit(
-    url: String,
-    permits: Int,
-    period: Duration = 1.seconds,
-): OkHttpClient.Builder = throw RuntimeException("Stub!")
-
-/**
- * An OkHttp interceptor that handles given url host's rate limiting.
- *
- * Examples:
- *
- * httpUrl = "https://api.anime.com".toHttpUrlOrNull(), permits = 5, period = 1.seconds =>  5 requests per second to api.anime.com
- * httpUrl = "https://imagecdn.anime.com".toHttpUrlOrNull(), permits = 10, period = 2.minutes  =>  10 requests per 2 minutes to imagecdn.anime.com
- *
- * @since extension-lib 16
- *
- * @param httpUrl [HttpUrl] The url host that this interceptor should handle. Will get url's host by using HttpUrl.host()
- * @param permits [Int]     Number of requests allowed within a period of units.
- * @param period [Duration] The limiting duration. Defaults to 1.seconds.
- */
-fun OkHttpClient.Builder.rateLimit(
-    httpUrl: HttpUrl,
-    permits: Int,
-    period: Duration = 1.seconds,
-): OkHttpClient.Builder = throw RuntimeException("Stub!")
-
-/**
- * An OkHttp interceptor that enforces conditional rate limiting based on a given condition.
- *
- * Examples:
- *
- * permits = 5, period = 1.seconds, shouldLimit = { it.host == "api.anime.com" } => 5 requests per second to api.anime.com.
- * permits = 10, period = 2.minutes, shouldLimit = { it.encodedPath.startsWith("/images/") } => 10 requests per 2 minutes to paths starting with "/images/".
- *
- * @since extension-lib 16
- *
- * @param permits [Int]     Number of requests allowed within a period of units.
- * @param period [Duration] The limiting duration. Defaults to 1.seconds.
- * @param shouldLimit       A predicate to determine whether the rate limit should apply to a given request.
- */
-fun OkHttpClient.Builder.rateLimit(
-    permits: Int,
-    period: Duration = 1.seconds,
-    shouldLimit: (HttpUrl) -> Boolean,
 ): OkHttpClient.Builder = throw RuntimeException("Stub!")
