+++ library/src/main/java/eu/kanade/tachiyomi/network/interceptor/SpecificHostRateLimitInterceptor.kt
@@ -1,10 +1,10 @@
+@file:Suppress("UnusedReceiverParameter")
+
 package eu.kanade.tachiyomi.network.interceptor
 
 import okhttp3.HttpUrl
 import okhttp3.OkHttpClient
 import java.util.concurrent.TimeUnit
-import kotlin.time.Duration
-import kotlin.time.Duration.Companion.seconds
 
 /**
  * An OkHttp interceptor that handles given url host's rate limiting.
@@ -14,8 +14,8 @@ import kotlin.time.Duration.Companion.seconds
  *
  * Examples:
  *
- * httpUrl = "https://api.anime.com".toHttpUrl(), permits = 5, period = 1, unit = seconds  =>  5 requests per second to api.anime.com
- * httpUrl = "https://cdn.thumbnails.com".toHttpUrl(), permits = 10, period = 2, unit = minutes  =>  10 requests per 2 minutes to cdn.thumbnails.com
+ * httpUrl = "https://api.anime.example".toHttpUrl(), permits = 5, period = 1, unit = seconds  =>  5 requests per second to api.anime.example
+ * httpUrl = "https://cdn.anime.example/images".toHttpUrl(), permits = 10, period = 2, unit = minutes  =>  10 requests per 2 minutes to cdn.anime.examples
  *
  * @since extension-lib 13
  *
@@ -24,8 +24,17 @@ import kotlin.time.Duration.Companion.seconds
  * @param period [Long]     The limiting duration. Defaults to 1.
  * @param unit [TimeUnit]   The unit of time for the period. Defaults to seconds.
  */
-@Suppress("unused_parameter", "DeprecatedCallableAddReplaceWith")
-@Deprecated("Use the version with kotlin.time [rateLimit] APIs instead.")
+@Deprecated(
+    message = "Use the version with kotlin.time APIs instead.",
+    replaceWith = ReplaceWith(
+        expression = "rateLimit(permits, period.toDuration(unit.toDurationUnit())) { it.host == httpUrl.host }",
+        imports = [
+            "mihonx.network.rateLimit",
+            "kotlin.time.toDuration",
+            "kotlin.time.toDurationUnit",
+        ]
+    )
+)
 fun OkHttpClient.Builder.rateLimitHost(
     httpUrl: HttpUrl,
     permits: Int,
